Create complement directory
