+++ server/src/services/utils.ts
@@ -1,41 +1,6 @@
-import {ComplementKind, ComplementScopeRegion} from "../compiler_analyzer/complementHint";
-import {Position} from "vscode-languageserver";
-import {SymbolScope} from "../compiler_analyzer/symbolScope";
 import {TextPosition} from "../compiler_tokenizer/textLocation";
 import {TokenObject} from "../compiler_tokenizer/tokenObject";
 import {SymbolObject} from "../compiler_analyzer/symbolObject";
-
-export function takeNarrowestHint(lhs: ComplementScopeRegion, rhs: ComplementScopeRegion): ComplementScopeRegion {
-    const lhsDiff = lhs.boundingLocation.getDifference();
-    const rhsDiff = rhs.boundingLocation.getDifference();
-
-    if (lhsDiff.line < rhsDiff.line) return lhs;
-    if (lhsDiff.line > rhsDiff.line) return rhs;
-    return lhsDiff.character < rhsDiff.character ? lhs : rhs;
-}
-
-/**
- * Find the scope that includes the specified position.
- */
-export function findScopeContainingPosition(scope: SymbolScope, caret: Position, path: string): SymbolScope {
-    const globalScope = scope.getGlobalScope();
-
-    let cursor: ComplementScopeRegion | undefined = undefined;
-    for (const hint of globalScope.completionHints) {
-        if (hint.complement !== ComplementKind.ScopeRegion) continue;
-
-        const location = hint.boundingLocation;
-        if (location.path !== path) continue;
-
-        if (location.positionInRange(caret)) {
-            cursor = cursor === undefined
-                ? hint
-                : takeNarrowestHint(cursor, hint);
-        }
-    }
-
-    return cursor?.targetScope ?? scope;
-}
 
 /**
  * Finds the token in the given list that contains the specified caret position.
